reposition in location controller
