+++ MapApp/Views/Locations/Recommend.cshtml
